requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for validation
pydantic==2.5.0
pydantic[email]==2.5.0

# HTTP client
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Azure OpenAI
openai==1.45.0

# PDF Generation
reportlab==4.0.7

# Visualizations
matplotlib==3.8.2
numpy==1.26.2

# Logging
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
